EOS-15472 : m0cp throwing error on IO
Resolved review comments for typecasting

+++ motr/io_pargrp.c
@@ -1347,7 +1347,7 @@ static int pargrp_iomap_paritybufs_alloc(struct pargrp_iomap *map)
 						     (i*row_per_seg)+ row][col];
 				data_buf_init(dbuf, ptr, obj_buffer_size(obj),
 					      PA_NONE);
-				(uint64_t)ptr += obj_buffer_size(obj);
+				(uintptr_t)ptr += obj_buffer_size(obj);
 
 				M0_LOG(M0_DEBUG, "row=%d col=%d dbuf=%p pbuf = %p ptr=%p",
 				       row, col, dbuf, pbuf, ptr);
